Healthcare Projects/6. Medical Chatbot/backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
pinecone
pypdf
tiktoken
langchain-pinecone
sentence-transformers